Update to evaluations entity
