+++ .vscode/settings.json
@@ -1,0 +1,22 @@
+{
+    "workbench.colorCustomizations": {
+        "activityBar.activeBackground": "#795ab2",
+        "activityBar.background": "#795ab2",
+        "activityBar.foreground": "#e7e7e7",
+        "activityBar.inactiveForeground": "#e7e7e799",
+        "activityBarBadge.background": "#c2957d",
+        "activityBarBadge.foreground": "#15202b",
+        "commandCenter.border": "#e7e7e799",
+        "sash.hoverBorder": "#795ab2",
+        "statusBar.background": "#614594",
+        "statusBar.foreground": "#e7e7e7",
+        "statusBarItem.hoverBackground": "#795ab2",
+        "statusBarItem.remoteBackground": "#614594",
+        "statusBarItem.remoteForeground": "#e7e7e7",
+        "titleBar.activeBackground": "#614594",
+        "titleBar.activeForeground": "#e7e7e7",
+        "titleBar.inactiveBackground": "#61459499",
+        "titleBar.inactiveForeground": "#e7e7e799"
+    },
+    "peacock.color": "#614594"
+}
