+++ src/Eras.Api/Controllers/EvaluationController.cs
@@ -1,11 +1,9 @@
-﻿using System.Xml.Linq;
-using Eras.Application.DTOs;
-using Eras.Application.Features.Evaluations.Commands.CreateEvaluation;
-using Eras.Application.Features.Students.Commands.CreateStudent;
+﻿using Eras.Application.DTOs;
+using Eras.Application.Features.Evaluations.Commands;
+using Eras.Application.Features.Evaluations.Queries;
 using Eras.Application.Models;
 using Eras.Domain.Entities;
 using MediatR;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 
 namespace Eras.Api.Controllers
@@ -47,6 +45,16 @@ namespace Eras.Api.Controllers
                     new { status = "error", message = "An error occurred during the evaluation creation process" }
                 );
             }
+        }
+
+        [HttpGet("summary")]
+        [ProducesResponseType(StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        public async Task<IActionResult> GetEvaluationProcessSumary()
+        {
+            _logger.LogInformation("Getting evaluation process summary");
+            GetEvaluationSummaryQuery summary = new();
+            return Ok(await _mediator.Send(summary));
         }
     }
 }
