--- src/Eras.Application/Features/EvaluationProcess/Queries/GetEvaluationProcessSumaryQueryHandler.cs
@@ -1,19 +1,0 @@
-﻿using Eras.Application.Contracts.Persistence;
-using Eras.Domain.Entities;
-using MediatR;
-using Microsoft.Extensions.Logging;
-
-namespace Eras.Application.Features.EvaluationProcess.Queries
-{
-    class GetEvaluationProcessSumaryQueryHandler
-        (IPollCohortRepository repository,
-        ILogger<GetEvaluationProcessSummaryQuery> logger)
-        : IRequestHandler<GetEvaluationProcessSummaryQuery, List<Poll>>
-    {
-        public async Task<List<Poll>> Handle(GetEvaluationProcessSummaryQuery request, CancellationToken cancellationToken)
-        {
-            var polls = await repository.GetAllAsync();
-            return [.. polls];
-        }
-    }
-}
