--- src/Eras.Application/Features/EvaluationProcess/Queries/GetEvaluationProcessSummaryQuery.cs
@@ -1,7 +1,0 @@
-﻿using MediatR;
-using Eras.Domain.Entities;
-
-namespace Eras.Application.Features.EvaluationProcess.Queries
-{
-    public class GetEvaluationProcessSummaryQuery: IRequest<List<Poll>>;
-}
