+++ src/Eras.Application/Features/Evaluations/Commands/CreateEvaluation/CreateEvaluationCommand.cs
@@ -1,10 +1,9 @@
-﻿using Eras.Application.Dtos;
-using Eras.Application.DTOs;
+﻿using Eras.Application.DTOs;
 using Eras.Application.Models;
 using Eras.Domain.Entities;
 using MediatR;
 
-namespace Eras.Application.Features.Evaluations.Commands.CreateEvaluation
+namespace Eras.Application.Features.Evaluations.Commands
 {
     public class CreateEvaluationCommand : IRequest<CreateComandResponse<Evaluation>>
     {
