+++ src/Eras.Application/Features/Evaluations/Queries/GetEvaluationSummaryQuery.cs
@@ -1,0 +1,7 @@
+﻿using MediatR;
+using Eras.Domain.Entities;
+
+namespace Eras.Application.Features.Evaluations.Queries
+{
+    public class GetEvaluationSummaryQuery: IRequest<List<Poll>>;
+}
